orc/requirements.txt
boto3
fastapi
mangum
python-jose[cryptography]
